+++ src/main/java/finalmission/reservation/domain/ReservationTime.java
@@ -1,4 +1,4 @@
-package finalmission.restaurant.domain;
+package finalmission.reservation.domain;
 
 
 import jakarta.persistence.Column;
