+++ src/main/java/finalmission/reservation/infrastructure/JpaReservationTimeRepository.java
@@ -1,6 +1,6 @@
-package finalmission.restaurant.infrastructure;
+package finalmission.reservation.infrastructure;
 
-import finalmission.restaurant.domain.ReservationTime;
+import finalmission.reservation.domain.ReservationTime;
 import java.time.LocalTime;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
